put setProperty into a separate file
